transparent-trust-ledger/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
SQLAlchemy==2.0.34
psycopg[binary]==3.2.12
pydantic==2.8.2
pydantic-settings==2.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
requests==2.32.3
razorpay==1.4.2